[feat,fix] 레이싱 경주 메인함수 작성 및 자동차 move 인터페이스로 구현
