+++ src/main/java/study/step3/Car.java
@@ -1,14 +1,4 @@
 package study.step3;
-
-public class Car {
-    private int position;
-    public Car(){
-        this.position = 1;
-    }
-    public int move(){
-        if(NumberChecker.checkCondition()){
-            position = position + 1;
-        }
-        return position;
-    }
+public interface Car{
+    int move();
 }
